refactor: api 수정
