+++ src/main/kotlin/com/example/mykku/dailymessage/DailyMessageController.kt
@@ -10,15 +10,17 @@ import org.springframework.format.annotation.DateTimeFormat
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
+import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
 import java.time.LocalDate
 
 @RestController
+@RequestMapping("/api/v1/daily-messages")
 class DailyMessageController(
     private val dailyMessageService: DailyMessageService
 ) {
-    @GetMapping("/api/v1/daily-messages")
+    @GetMapping
     fun getDailyMessages(
         @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) date: LocalDate,
         @RequestParam(defaultValue = "0") page: Int,
@@ -40,7 +42,7 @@ class DailyMessageController(
         )
     }
 
-    @GetMapping("/api/v1/daily-message/{id}")
+    @GetMapping("/{id}")
     fun getDailyMessage(
         @PathVariable id: Long
     ): ResponseEntity<ApiResponse<DailyMessageResponse>> {
