--- src/main/kotlin/com/example/mykku/like/dto/LikeBoardRequest.kt
@@ -1,5 +1,0 @@
-package com.example.mykku.like.dto
-
-data class LikeBoardRequest(
-    val boardId: Long
-)
