--- src/main/kotlin/com/example/mykku/like/dto/LikeDailyMessageCommentRequest.kt
@@ -1,5 +1,0 @@
-package com.example.mykku.like.dto
-
-data class LikeDailyMessageCommentRequest(
-    val dailyMessageCommentId: Long
-)
