--- src/main/kotlin/com/example/mykku/like/dto/LikeFeedCommentRequest.kt
@@ -1,5 +1,0 @@
-package com.example.mykku.like.dto
-
-data class LikeFeedCommentRequest(
-    val feedCommentId: Long
-)
