--- src/main/kotlin/com/example/mykku/like/dto/LikeFeedRequest.kt
@@ -1,5 +1,0 @@
-package com.example.mykku.like.dto
-
-data class LikeFeedRequest(
-    val feedId: Long,
-)
