+++ src/test/kotlin/com/example/mykku/like/LikeControllerTest.kt
@@ -2,7 +2,6 @@ package com.example.mykku.like
 
 import com.example.mykku.BaseControllerTest
 import com.example.mykku.board.domain.Board
-import com.example.mykku.like.dto.LikeBoardRequest
 import com.example.mykku.member.domain.Member
 import com.example.mykku.member.domain.SocialProvider
 import com.example.mykku.util.TestTokenGenerator
@@ -37,7 +36,7 @@ class LikeControllerTest : BaseControllerTest() {
         RestAssured.given()
             .header("Authorization", authHeader)
             .`when`()
-            .get("/api/v1/boards/like")
+            .get("/api/v1/likes/boards")
             .then()
             .statusCode(200)
             .body("message", equalTo("즐겨찾기한 게시판 목록을 성공적으로 조회하였습니다."))
@@ -50,7 +49,7 @@ class LikeControllerTest : BaseControllerTest() {
         // when & then
         RestAssured.given()
             .`when`()
-            .get("/api/v1/boards/like")
+            .get("/api/v1/likes/boards")
             .then()
             .statusCode(401)
     }
@@ -77,15 +76,13 @@ class LikeControllerTest : BaseControllerTest() {
             )
         )
         val authHeader = TestTokenGenerator.getBearerToken("member1")
-        val request = LikeBoardRequest(boardId = board.id!!)
 
         // when & then
         RestAssured.given()
             .header("Authorization", authHeader)
             .contentType(ContentType.JSON)
-            .body(request)
             .`when`()
-            .post("/api/v1/board/like")
+            .post("/api/v1/likes/boards/{boardId}", board.id!!)
             .then()
             .statusCode(200)
             .body("message", equalTo("게시판 즐겨찾기가 성공적으로 처리되었습니다."))
@@ -96,15 +93,11 @@ class LikeControllerTest : BaseControllerTest() {
     @Test
     @DisplayName("게시판 좋아요 - 인증되지 않은 사용자")
     fun `likeBoard - 인증되지 않은 사용자는 좋아요할 수 없다`() {
-        // given
-        val request = LikeBoardRequest(boardId = 1L)
-
         // when & then
         RestAssured.given()
             .contentType(ContentType.JSON)
-            .body(request)
             .`when`()
-            .post("/api/v1/board/like")
+            .post("/api/v1/likes/boards/{boardId}", 1L)
             .then()
             .statusCode(401)
     }
